+++ Kuroba/app/src/main/java/com/github/adamantcheese/chan/core/site/sites/Lainchan.java
@@ -59,7 +59,7 @@ public class Lainchan
         }
 
         @Override
-        public String desktopUrl(Loadable loadable, int postNo) {
+        public String desktopUrl(Loadable loadable, long postNo) {
             if (loadable.isCatalogMode()) {
                 return getUrl().newBuilder().addPathSegment(loadable.boardCode).toString();
             } else if (loadable.isThreadMode()) {
