+++ Kuroba/app/src/main/java/com/github/adamantcheese/chan/core/site/sites/dvach/Dvach.java
@@ -62,7 +62,7 @@ public class Dvach
         }
 
         @Override
-        public String desktopUrl(Loadable loadable, int postNo) {
+        public String desktopUrl(Loadable loadable, long postNo) {
             if (loadable.isCatalogMode()) {
                 return getUrl().newBuilder().addPathSegment(loadable.boardCode).toString();
             } else if (loadable.isThreadMode()) {
